book::rpc: Nuke
* mv models into `mod crate::db::models`
